printController: Do not allow selection of hidden model fields
Do not allow users to select hidden model fields for display in
data-grids.

This is useful where a model field is not present in the collection
database 'view'. If sucha field is selected for display, it is added
to the SQL select statement and causes an error.

In a model definition:
    $this->setHidden('fieldname')