Make the back button work more naturally when returning from ajaxed, related items.
I.e., reload the page to display the main view when hitting the back button from
a related items view.

If we don't do this then the back button will go back too far, from the users viewpoint.
Obviously, using the forward button won't return to the ajaxed related items. To achieve
that, we would need to adjust uzERP to respond to a url containing a 'something_related'
hash location, generating the page to be viewed.