PersonController: after save, pass person_id and slmaster_id to response
When adding a new person from a	new sales order	or invoice
the current record is reloaded	to update the form inputs.
We need	to pass	through the slmaster and person	ids to
allow the form to display the valid company and	person.

Also fixed/added:

- Set the person module javascript in a smarty variable so it
can be used in the dialog to drive form	behaviour.

- Fix method dignatures	that viaolate LSP.

- Remove some unused code.