Javascript, rules.js: add parameters to URL on refresh after dialog.
Allows a controller action to add parameters to a sendTo() call
and have thos paramaters added to the target URL for the
target action to use.

- Remove 'save and add another' button on 'add' type dialogs
as it isn't needed or desirable.

- Add ability to load module JS from another module
if a controller action has asked for it. Some form dialogs
will have behavior that requires JS, e.g. adding a person
when adding a new sales order requires the Persons
module JS.