Smarty function.input.php: For dialogs, don't clobber a value already set.
When there is a collision between the controller model (e.g. sorder) and
model used in the dialog (e.g. the party model), both with fields with
the same name (type in this example), we want to keep the value already set
not replace it with the parent (sorder) model value.

Also:

- Add setting of error classes and message on matched fields.

- Output required field marker with span tag and class for further styling.