SOrderLine model: switch from PriceFormatter to CurrencyFormatter
PriceFormatter causes XML output of the ProForma invoice
to fail becuase it returns the &pound; html entity instead
insted of '£'. Price Formatter is also not currency aware.

Tech note:

&pound; is later converted to &#38;pound; when building
the XML resulting in '&pound;' on the PDF output.

The CurrencyFormatter retrieves the correct unicode symbol,
which is converted to &#163; in the XML and shows as '£'
on the PDF output.