fix(slcustomer): improve performance of SLCustomer display
On display, ::view() checks for current activity for the customer
to determine if they can be made inactive.

It uses ::hasCurrentActivity() on the SLCustomer model,
which uses an un-constrained query on sales orders,
invoices and ledger transactions. A large number of records
can be returned making things slow and sometimes exceeds
PHP's memory limit. We only need to know if at least one
record exists.

This fix uses SearchHandler::setLimit(1) to limit the query
result.