fix(manufacturing): Remove date from cost sheet search
Cost sheets for a specific date cannot be produced using the
current method. For example, searching mf_structures by start/end
date is not a valid way to calaculate costs at a particular date.
uzERP stores costs on mf_structures that the cost sheet uses,
but these will always be from the most recent recost.

Remove the date because it just confuses users.